fix: Stop code running multiple times per morsel
The viewer-player.loadingComplete() function is called constantly
during video playback. This causes code syntax highlighting and the
Downloadables section move to run many times per video morsel.

This commit is a workaround instead of long term solution. A jQuery
data() record is set on each processed DOM item. Items with an existing
data() value are not processed.

Long term solution is to determine which Angular Directive(s) or
Controller(s) needs to be extended to apply these changes without
the data() or setTimeout().